Move unit test
RISDEV-4432
